Prepare custom PSN setup
The Verify cluster needs to have a VPC endpoint available for the teams
to use PSN. This isn't standard core GSP requirement, therefore the
snowflake is to be living with the verify cluster config.

We will need to apply the module. Instead of just going into this module
directory, we will be calling the module from different terraform file,
which should allow us to nicely separate the potentially sensitive data
from the "logic".

Co-authored-by: Rafal Proszowski <rafal.proszowski@digital.cabinet-office.gov.uk>
Co-authored-by: Issy Long <isabell.long@digital.cabinet-office.gov.uk>
Co-authored-by: Daniel Blair <daniel.blair@digital.cabinet-office.gov.uk>