Limit number of Subnet IDs for VPC endpoint
As it turns out, Cloud Fundamentals the company providing PSNaaS has set
us up with two AZs available for our VPC endpoint. We are running the
GSP cluster across three AZs, meaning we have to do some work to pick
the correct two AZs (`2a` and `2b`) from the Cluster state.

This is very fragile, and has no guarantees to be indefinitely
working... We could be looking at validating the AZs across the
accounts, as AWS likes to mix the names for the AZs up a little.

For further reading:

https://docs.aws.amazon.com/ram/latest/userguide/working-with-az-ids.html
https://www.terraform.io/docs/providers/aws/d/availability_zone.html
https://www.terraform.io/docs/providers/aws/d/subnet_ids.html

Co-authored-by: Rafal Proszowski <rafal.proszowski@digital.cabinet-office.gov.uk>
Co-authored-by: Issy Long <isabell.long@digital.cabinet-office.gov.uk>
Co-authored-by: Daniel Blair <daniel.blair@digital.cabinet-office.gov.uk>