Define private hosted zone
We believe that the endpoint could be sensitive piece of information. We
don't feel comfortable exposing it to the wild with the route53
record... Although, we trust that the provider will only approve
connections based on the data we provide.

Instead, we can setup a private hosted zone, for our VPC and work with
that. There is no reason, the private hosted zone couldn't live with the
rest of GSP core. In order to unblock ourselves, we're willing to have
it as part of the snowflake.