Give the PSN VPC endpoint its own security group
Currently, the PSN VPC endpoint is in the "worker" security group,
which is good to get us going but is possibly too permissive.  We only
need access to a single port on the VPC endpoint.

This commit gives the VPC endpoint its own security group and grants
access from the worker security group to the VPC endpoint security
group on port 3128.