Add more workers and resize the boxes
We are experiencing issues with Concourse in particular. Jobs are slow,
things are not being scheduled properly or in timely fashion.

Adding workers, should ensure spread of the load, and resizing the
instances should provide it with more compute power to perform certain
jobs faster.

One thing to watch out for, is that we believe the concourse workers,
which are in fact containers running on the nodes are magically sizing
up to the fullest potential. We should perhaps look into if that's true
and figure out if the CI AWS instances are being fully used at the peak
times.