update egress (ServiceEntry) configuration
we now use the istio ServiceEntry format to describe what services
external to the cluster we allow egress to so that we have full control
over the configuration of ports etc enabling whitelisting non-http
services.

The available options for the `service` block can be found in the Istio
documentation: https://istio.io/docs/reference/config/networking/v1alpha3/service-entry/#ServiceEntry

in addition to reformatting the existing entries, we also add a new
entry for connecting to postgres (rds) services within the eu-west-2
region.