use resolution:NONE for postgres ServiceEntry
it's not possible to use a wildcard host with DNS resolution.

using resolution:NONE means that the destination IP is trusted and the
istio sidecar will not attempt to lookup the IP itself.

using resolution:NONE alone would introduce the possibility to
exfiltrate data by spoofing SNI, so we scope the addresses to only
thoses within the VPC.