custom domains: Try removing forwarding of headers
In experimenting with my CloudFront-distributed-sandbox-canary it seems turning
this off allows CloudFront to connect to our GSP origins. I suspect this is
becuase with it on, CloudFront will attempt to use the specified Host header,
and then get upset when GSP serves a certificate that does not match that Host.